Don't build Python on OSX for Travis
Travis docs declare a lack of support for Python-on-OSX builds (bummer!), but at least the looper documentation landing page explains the OSX setup. There are some proposed workarounds for the apparent disdain of Travis for Python-on-OSX, and we could try one of those or wait until team Travis incorporates more robust support for this, but for now only do the (much faster and more straightforward) Linux build.